Refractoring done and minor fix
